apps/backend/pyproject.toml
[tool.poetry]
name = "travelbot-backend"
version = "0.1.0"
description = ""
authors = ["Herbstblatt <48344463+Herbstblatt@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
asyncpg = "^0.29.0"
python-jose = "^3.3.0"
aiogram = "^3.4.1"
tortoise-orm = "^0.20.0"
python-multipart = "^0.0.9"
aerich = "^0.7.2"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
